backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]>=2.0.25
asyncpg
pydantic[email]>=2.12.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
requests==2.31.0
cryptography>=41.0.0
psycopg2-binary